requirements.txt
# Railway deployment requirements (CPU-only for faster builds)
numpy>=1.21.0
Pillow>=8.0.0

# PyTorch CPU-only (smaller, faster install)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# Web Application
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
gunicorn>=21.0.0
